Added code for petri-visual team
Petrinet how has a "Petridrawer",
which has a draw function, that takes the Petrinet and returns a
buffered image.

Petridrawer has a zoom function, where the size of the objects can be
chosen

If another size of the window is preferred, then change it directly in
the PetriWindow class. Change values width and height. (cannot be done
using a constructor)